+++ mtpy/gui/modem_mesh_builder.py
@@ -110,7 +110,7 @@ class ModEM_Mesh_Window(QtGui.QMainWindow):
         fn = str(fn_dialog.getOpenFileName(caption='Choose ModEM model file',
                                        filter='*.rho'))
                                        
-        self.mesh_widget.model_obj = modem.Model()
+        self.mesh_widget.model_obj = mtpy.modeling.ModEM.Model()
         self.mesh_widget.model_obj.read_model_file(fn)
 
         self.mesh_widget.dir_path = os.path.dirname(fn)
@@ -139,7 +139,7 @@ class MeshWidget(QtGui.QWidget):
     
     def __init__(self):
         super(MeshWidget, self).__init__()
-        self.model_obj = modem.Model()
+        self.model_obj = mtpy.modeling.ModEM.Model()
         self.mpl_widget = MeshPlot()
         
         #sys.stdout = MyStream()
@@ -357,7 +357,7 @@ class MeshWidget(QtGui.QWidget):
         fn = str(fn_dialog.getOpenFileName(caption='Choose ModEM model file',
                                        filter='*.rho'))
                                        
-        self.model_obj = modem.Model()
+        self.model_obj = mtpy.modeling.ModEM.Model()
         self.model_obj.read_model_file(fn)
 
         self.dir_path = os.path.dirname(fn)
@@ -416,7 +416,7 @@ class MeshWidget(QtGui.QWidget):
         for fn in fn_list:
             self.edi_list.append(str(fn))
         
-        self.model_obj = modem.Model(edi_list=self.edi_list)
+        self.model_obj = mtpy.modeling.ModEM.Model(edi_list=self.edi_list)
         
         self.model_obj.get_station_locations()
