+++ mtpy/modeling/ModEM/__init__.py
@@ -1,5 +1,6 @@
 from .exception import ModEMError, DataError
 from .station import Stations
 from .data import Data
+from .model import Model
 
-__all__ = ['ModEMError', 'DataError', 'Stations', 'Data']
+__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model']
