+++ mtpy/utils/array2raster.py
@@ -4,6 +4,8 @@ Created on Sun May 11 12:15:37 2014
 
 @author: jrpeacock
 """
+import mtpy.modeling.ModEM
+
 try:
     from osgeo import ogr, gdal, osr
 except ImportError:
@@ -52,7 +54,7 @@ class ModEM_to_Raster(object):
         get model to put into array
         """
 
-        model_obj = modem.Model()
+        model_obj = mtpy.modeling.ModEM.Model()
         model_obj.model_fn = self.model_fn
         model_obj.read_model_file()
         
@@ -82,7 +84,7 @@ class ModEM_to_Raster(object):
         self.pad_east = pad_east
         self.pad_north = pad_north
             
-        model_obj = modem.Model()
+        model_obj = mtpy.modeling.ModEM.Model()
         model_obj.model_fn = self.model_fn
         model_obj.read_model_file()
         
@@ -114,7 +116,7 @@ class ModEM_to_Raster(object):
         
         """
         
-        model_obj = modem.Model()
+        model_obj = mtpy.modeling.ModEM.Model()
         model_obj.model_fn = self.model_fn
         model_obj.read_model_file()
